Move AppStyles to theme package
