+++ src/main/kotlin/org/wycliffeassociates/otter/jvm/app/theme/AppColors.kt
@@ -5,8 +5,6 @@ import tornadofx.c
 
 abstract class AppColors {
     val white: Color = Color.WHITE
-    val lightGray: Color = c("#DFDEE3")
-    val gray: Color = c("#AAA")
     val appRed: Color = c("#CC4141")
     val appBlue: Color = c("#0094F0")
     val appGreen: Color = c("#58BD2F")
@@ -14,6 +12,7 @@ abstract class AppColors {
     abstract val base: Color
     abstract val defaultBackground: Color
     abstract val defaultText: Color
+    abstract val subtitle: Color
     abstract val cardBackground: Color
     abstract val disabledCardBackground: Color
     abstract val colorlessButton: Color
